remove repetitions
