+++ lib/api/mailboxes.js
@@ -6,7 +6,7 @@ const imapTools = require('../../imap-core/lib/imap-tools');
 const tools = require('../tools');
 const roles = require('../roles');
 const util = require('util');
-const { sessSchema, sessIPSchema, booleanSchema } = require('../schemas');
+const { sessSchema, sessIPSchema, booleanSchema, mailboxPathValidator } = require('../schemas');
 const { userId, mailboxId } = require('../schemas/request/general-schemas');
 const { successRes } = require('../schemas/response/general-schemas');
 const { GetMailboxesResult } = require('../schemas/response/mailboxes-schemas');
@@ -283,23 +283,7 @@ module.exports = (db, server, mailboxHandler) => {
                     path: Joi.string()
                         .regex(/\/{2,}|\/$/, { invert: true })
                         .max(MAX_MAILBOX_NAME_LENGTH * MAX_SUB_MAILBOXES + 127)
-                        .custom((value, helpers) => {
-                            const splittedPathArr = value.split('/');
-
-                            if (splittedPathArr.length > MAX_SUB_MAILBOXES) {
-                                // too many paths
-                                return helpers.message(`The mailbox path cannot be more than ${MAX_SUB_MAILBOXES} levels deep`);
-                            }
-
-                            for (const pathPart of splittedPathArr) {
-                                if (pathPart.length > MAX_MAILBOX_NAME_LENGTH) {
-                                    // part too long error
-                                    return helpers.message(`Any part of the mailbox path cannot be longer than ${MAX_MAILBOX_NAME_LENGTH} chars long`);
-                                }
-                            }
-
-                            return value;
-                        }, 'Mailbox path validation')
+                        .custom(mailboxPathValidator, 'Mailbox path validation')
                         .required()
                         .description('Full path of the mailbox, folders are separated by slashes, ends with the mailbox name (unicode string)'),
                     hidden: booleanSchema.default(false).description('Is the folder hidden or not. Hidden folders can not be opened in IMAP.'),
@@ -559,23 +543,7 @@ module.exports = (db, server, mailboxHandler) => {
                     path: Joi.string()
                         .regex(/\/{2,}|\/$/, { invert: true })
                         .max(MAX_MAILBOX_NAME_LENGTH * MAX_SUB_MAILBOXES + 127)
-                        .custom((value, helpers) => {
-                            const splittedPathArr = value.split('/');
-
-                            if (splittedPathArr.length > MAX_SUB_MAILBOXES) {
-                                // too many paths
-                                return helpers.message(`The mailbox path cannot be more than ${MAX_SUB_MAILBOXES} levels deep`);
-                            }
-
-                            for (const pathPart of splittedPathArr) {
-                                if (pathPart.length > MAX_MAILBOX_NAME_LENGTH) {
-                                    // part too long error
-                                    return helpers.message(`Any part of the mailbox path cannot be longer than ${MAX_MAILBOX_NAME_LENGTH} chars long`);
-                                }
-                            }
-
-                            return value;
-                        }, 'Mailbox path validation')
+                        .custom(mailboxPathValidator, 'Mailbox path validation')
                         .description('Full path of the mailbox, use this to rename an existing Mailbox'),
                     retention: Joi.number()
                         .empty('')
