+++ lib/schemas.js
@@ -2,6 +2,7 @@
 
 const EJSON = require('mongodb-extended-json');
 const Joi = require('joi');
+const { MAX_SUB_MAILBOXES, MAX_MAILBOX_NAME_LENGTH } = require('./consts');
 
 const sessSchema = Joi.string().max(255).label('Session identifier').description('Session identifier for the logs');
 const sessIPSchema = Joi.string()
@@ -82,6 +83,24 @@ const metaDataValidator = () => (value, helpers) => {
     return strValue;
 };
 
+const mailboxPathValidator = (value, helpers) => {
+    const splittedPathArr = value.split('/');
+
+    if (splittedPathArr.length > MAX_SUB_MAILBOXES) {
+        // too many paths
+        return helpers.message(`The mailbox path cannot be more than ${MAX_SUB_MAILBOXES} levels deep`);
+    }
+
+    for (const pathPart of splittedPathArr) {
+        if (pathPart.length > MAX_MAILBOX_NAME_LENGTH) {
+            // part too long error
+            return helpers.message(`Any part of the mailbox path cannot be longer than ${MAX_MAILBOX_NAME_LENGTH} chars long`);
+        }
+    }
+
+    return value;
+};
+
 const mongoCursorSchema = Joi.string().trim().empty('').custom(mongoCursorValidator({}), 'Cursor validation').max(1024);
 const pageLimitSchema = Joi.number().default(20).min(1).max(250).label('Page size');
 const pageNrSchema = Joi.number().default(1).label('Page number').description('Current page number. Informational only, page numbers start from 1');
@@ -111,5 +130,6 @@ module.exports = {
     pageLimitSchema,
     booleanSchema,
     metaDataSchema,
-    usernameSchema
+    usernameSchema,
+    mailboxPathValidator
 };
